requirements.txt
flask
flask-cors
pyjwt
werkzeug
pyyaml
sqlalchemy
opencv-python
